refactor: last changes in migrations and models

+++ db/migrate/20240430111519_change_money_columns.rb
@@ -7,11 +7,11 @@ class ChangeMoneyColumns < ActiveRecord::Migration[7.1]
     remove_column :trades, :final_amount_cents
     remove_column :trades, :final_amount_currency
 
-    add_column :blockchain_payments, :amount_cents, :float
+    add_column :blockchain_payments, :amount_cents, :bigint
     add_column :blockchain_payments, :amount_currency, :string
-    add_column :trades, :original_amount_cents, :float
+    add_column :trades, :original_amount_cents, :bigint
     add_column :trades, :original_amount_currency, :string
-    add_column :trades, :final_amount_cents, :float
+    add_column :trades, :final_amount_cents, :bigint
     add_column :trades, :final_amount_currency, :string
   end
 end
